feat(boostpowjob): add scriptVersion and useACICBoost methods
We need to be able to know the script verison, which has to do with whether we can use ASICBoost and
also with extraNonce2Size. We need to know that to mine boost. I have added functions for those with
tests.

Additionally, the jenky formatting in boost-pow-job-model.ts has been fixed. I also removed graph-search-query.ts

Update version to 1.4.1